requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
charset-normalizer==2.1.1
discord.py==2.0.1
Django==4.1.2
django-annoying==0.10.6
frozenlist==1.3.1
idna==3.4
multidict==6.0.2
pycodestyle==2.9.1
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.4
wakeonlan==2.1.0
yarl==1.8.1